Restructure data hierarchy
