+++ src/data/upload/house/uploadHouseBillAuthorships.ts
@@ -5,8 +5,8 @@ import {
   PrismaClient,
 } from "@prisma/client";
 import expect from "expect";
-import parseBillAuthorships from "../scraper/congressBillAuthorships/billAuthorshipParser";
-import { getRepresentativeUrls } from "../scraper/representatives/listRepresentatives";
+import parseBillAuthorships from "../../scraper/house/billAuthorship/billAuthorshipParser";
+import { getRepresentativeUrls } from "../../scraper/house/members/listRepresentatives";
 
 type ParsedBillAuthorship = Omit<BillAuthorship, "id">;
